Split out internal resolver tests
